Fix: comparison of narrow type to wide type in loop (potential for infinite loops)

+++ src/linkgraph/linkgraph.h
@@ -495,7 +495,7 @@ public:
 	 * Get the current size of the component.
 	 * @return Size.
 	 */
-	inline uint Size() const { return (uint)this->nodes.size(); }
+	inline uint16 Size() const { return (uint16)this->nodes.size(); }
 
 	/**
 	 * Get date of last compression.
